+++ .ci-scripts/x86-64-unknown-linux-gnu-release.bash
@@ -9,8 +9,6 @@ then
   exit 1
 fi
 
-TODAY=$(date +%Y%m%d)
-
 # Compiler target parameters
 ARCH=x86-64
 PIC=true
@@ -24,16 +22,15 @@ TRIPLE=${ARCH}-${VENDOR}-${OS}
 MAKE_PARALLELISM=4
 BUILD_PREFIX=$(mktemp -d)
 DESTINATION=${BUILD_PREFIX}/lib/pony
-PONY_VERSION="nightly-${TODAY}"
 
 # Asset information
 PACKAGE_DIR=$(mktemp -d)
 PACKAGE=ponyc-${TRIPLE}
 
 # Cloudsmith configuration
-CLOUDSMITH_VERSION=${TODAY}
+CLOUDSMITH_VERSION=$(cat VERSION)
 ASSET_OWNER=ponylang
-ASSET_REPO=nightlies
+ASSET_REPO=releases
 ASSET_PATH=${ASSET_OWNER}/${ASSET_REPO}
 ASSET_FILE=${PACKAGE_DIR}/${PACKAGE}.tar.gz
 ASSET_SUMMARY="Pony compiler"
@@ -42,8 +39,7 @@ ASSET_DESCRIPTION="https://github.com/ponylang/ponyc"
 # Build pony installation
 echo "Building ponyc installation..."
 make install prefix=${BUILD_PREFIX} default_pic=${PIC} arch=${ARCH} \
-  -j${MAKE_PARALLELISM} -f Makefile-lib-llvm symlink=no \
-  version="${PONY_VERSION}"
+  -j${MAKE_PARALLELISM} -f Makefile-lib-llvm symlink=no
 
 # Package it all up
 echo "Creating .tar.gz of ponyc installation..."