+++ src/libponyc/pass/pass.h
@@ -297,6 +297,9 @@ typedef struct pass_opt_t
   char* link_arch;
   char* linker;
   char* link_ldcmd;
+#ifndef NDEBUG
+  const char* llvm_args;
+#endif
 
   char* triple;
   char* cpu;