+++ test/libponyrt/util.cc
@@ -1,5 +1,11 @@
 #include <gtest/gtest.h>
+#include "util.h"
 
+void Main_runtime_override_defaults_oo(void* opt)
+{
+  (void)opt;
+  return;
+}
 
 int main(int argc, char** argv)
 {